Move `getOuterEditor` near the EmbeddedCodeEditorWidget
