+++ src/vs/editor/contrib/gotoSymbol/browser/peek/referencesController.ts
@@ -14,7 +14,8 @@ import { Position } from '../../../../common/core/position.js';
 import { Range } from '../../../../common/core/range.js';
 import { IEditorContribution } from '../../../../common/editorCommon.js';
 import { Location } from '../../../../common/languages.js';
-import { getOuterEditor, PeekContext } from '../../../peekView/browser/peekView.js';
+import { PeekContext } from '../../../peekView/browser/peekView.js';
+import { getOuterEditor } from '../../../../browser/widget/codeEditor/embeddedCodeEditorWidget.js';
 import * as nls from '../../../../../nls.js';
 import { CommandsRegistry } from '../../../../../platform/commands/common/commands.js';
 import { IConfigurationService } from '../../../../../platform/configuration/common/configuration.js';
